+++ src/HttpClientRequest.php
@@ -49,10 +49,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.0.0
      *
      * @param string $header The header.
+     *
+     * @return $this
      */
-    public function addHeader(string $header): void
+    public function addHeader(string $header): self
     {
         $this->headers[] = $header;
+
+        return $this;
     }
 
     /**
@@ -193,10 +197,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.1.0
      *
      * @param FilePathInterface $caCertificate The certificate authority (CA) certificate path.
+     *
+     * @return $this
      */
-    public function setCACertificate(FilePathInterface $caCertificate): void
+    public function setCACertificate(FilePathInterface $caCertificate): self
     {
         $this->caCertificate = $caCertificate;
+
+        return $this;
     }
 
     /**
@@ -205,10 +213,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.1.0
      *
      * @param FilePathInterface $clientCertificate The client certificate path.
+     *
+     * @return $this
      */
-    public function setClientCertificate(FilePathInterface $clientCertificate): void
+    public function setClientCertificate(FilePathInterface $clientCertificate): self
     {
         $this->clientCertificate = $clientCertificate;
+
+        return $this;
     }
 
     /**
@@ -217,10 +229,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.3.0
      *
      * @param string $clientCertificatePassword The client certificate password.
+     *
+     * @return $this
      */
-    public function setClientCertificatePassword(string $clientCertificatePassword): void
+    public function setClientCertificatePassword(string $clientCertificatePassword): self
     {
         $this->clientCertificatePassword = $clientCertificatePassword;
+
+        return $this;
     }
 
     /**
@@ -229,10 +245,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.3.0
      *
      * @param string $clientCertificateType
+     *
+     * return $this
      */
-    public function setClientCertificateType(string $clientCertificateType): void
+    public function setClientCertificateType(string $clientCertificateType): self
     {
         $this->clientCertificateType = $clientCertificateType;
+
+        return $this;
     }
 
     /**
@@ -241,10 +261,14 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.1.0
      *
      * @param FilePathInterface $clientKey The client key path.
+     *
+     * @return $this
      */
-    public function setClientKey(FilePathInterface $clientKey): void
+    public function setClientKey(FilePathInterface $clientKey): self
     {
         $this->clientKey = $clientKey;
+
+        return $this;
     }
 
     /**
@@ -254,11 +278,15 @@ class HttpClientRequest implements HttpClientRequestInterface
      *
      * @param string            $name     The name.
      * @param FilePathInterface $filePath The file path.
+     *
+     * @return $this
      */
-    public function setFile(string $name, FilePathInterface $filePath): void
+    public function setFile(string $name, FilePathInterface $filePath): self
     {
         $this->rawContent = '';
         $this->files[$name] = $filePath;
+
+        return $this;
     }
 
     /**
@@ -268,11 +296,15 @@ class HttpClientRequest implements HttpClientRequestInterface
      *
      * @param string $name  The name.
      * @param string $value The value.
+     *
+     * @return $this
      */
-    public function setPostField(string $name, string $value): void
+    public function setPostField(string $name, string $value): self
     {
         $this->rawContent = '';
         $this->postFields[$name] = $value;
+
+        return $this;
     }
 
     /**
@@ -281,12 +313,16 @@ class HttpClientRequest implements HttpClientRequestInterface
      * @since 1.0.0
      *
      * @param string $rawContent The raw content.
+     *
+     * @return $this
      */
-    public function setRawContent(string $rawContent): void
+    public function setRawContent(string $rawContent): self
     {
         $this->postFields = [];
         $this->files = [];
         $this->rawContent = $rawContent;
+
+        return $this;
     }
 
     /**