+++ src/HttpClientResponseInterface.php
@@ -23,8 +23,10 @@ interface HttpClientResponseInterface
      * @since 1.0.0
      *
      * @param string $header The header.
+     *
+     * @return $this
      */
-    public function addHeader(string $header): void;
+    public function addHeader(string $header): self;
 
     /**
      * Returns the content.